Expose the unpublishing information through GetContentCollection
Adds a new supported field "unpublishing" to
GetContentCollection. If the content item is unpublished, this will
contain the type, explanation, alternative_path and unpublished_at
information.

This will enable publishing applications using the Publishing API to
get back the information they put in when they unpublish content, and
was motivated by this requirement for the Maslow application.