Monkey patch ActiveRecord to kill ContentItem class
These monkey patches are used so that we can continue using a
polymorphic association on LockVersion to ContentItem/Edition without
changing what is saved in the database currently - namely a
`target_type` of "ContentItem".

This allows us to swap from this branch to the currently released
version without pain.

This patching should live for a very short period of time as we will not
need to maintain backwards compatibility once this is deployed and we
also intend to drop the lock_version table.